linked users with posts

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const fs = require('fs');
 const Path = require('path');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -56,22 +57,32 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     // console.log('In Create Post');
     // console.log(imageUrl);
+    let creator;
 
     const post = new Post({
         title: title,
         content: content,
-        creator: {
-            name: 'Soumyanil'
-        },
+        creator: req.userId,
         imageUrl: imageUrl
     });
 
     post.save()
         .then(result => {
-            // console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: "Post created successfully!",
-                post: result,
+                post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             });
         })
         .catch(err => {
@@ -117,6 +128,11 @@ exports.editPost = (req, res, next) => {
         //     errors: errors
         // });
 
+    }
+    if (post.creator.toStrng() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
     }
     const postId = req.params.postId;
     const title = req.body.title;
@@ -173,12 +189,23 @@ exports.deletePost = (req, res, next) => {
                 throw error;
             }
             // Check if the logged in user created the post
+            if (post.creator.toStrng() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
 
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            // console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             res.status(200).json({
                 message: 'Post successfully deleted!'
             })
